Move ActorViewHolder.kt into ActorsAdapter.kt Move MovieViewHolder.kt into MoviesAdapter.kt
