+++ app/src/main/java/ru/aevd/androidacademymovieapp/ActorsAdapter.kt
@@ -3,7 +3,10 @@ package ru.aevd.androidacademymovieapp
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.ImageView
+import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import com.bumptech.glide.Glide
 import ru.aevd.androidacademymovieapp.data.Actor
 
 class ActorsAdapter: RecyclerView.Adapter<ActorViewHolder>() {
@@ -26,3 +29,19 @@ class ActorsAdapter: RecyclerView.Adapter<ActorViewHolder>() {
     }
 
 }
+
+class ActorViewHolder(itemView: View):  RecyclerView.ViewHolder(itemView) {
+    private val imgActor = itemView.findViewById<ImageView>(R.id.img_actor)
+    private val actorName = itemView.findViewById<TextView>(R.id.actor_name)
+
+    fun onBind(actor: Actor) {
+        actorName.text = actor.name
+
+        Glide.with(context)
+                .load(actor.picture)
+                .into(imgActor)
+    }
+}
+
+private val RecyclerView.ViewHolder.context
+    get() = this.itemView.context
